requirements.txt
selenium==4.17.2
webdriver-manager==4.0.1
requests==2.31.0
pandas==2.2.0
openpyxl==3.1.2
xlwt==1.3.0
psutil==5.9.8
pillow==10.2.0
pywin32==306
pyinstaller==6.3.0
selenium
webdriver_manager
pandas
openpyxl
xlwt
requests
colorama
psutil
Pillow
pywin32
pyinstaller 